lib::wavelet; move testing gear into its own module
too much pollution around the tests.  going to work on the public / private
split.  especially tests for tests needed to go.  tests we don't want in CI will
likely be making the journey.
